# docs/swagger.yaml
definitions:
...
  models.CalculateExamGrades:
    properties:
      exam_id:
        type: integer
      student_id:
        type: integer
    type: object
...
  models.CreateDiscussionRequest:
    properties:
      description:
        type: string
      replies: {}
      student_id:
        type: integer
      topic:
        type: string
    type: object
  models.CreateExamAnswersRequest:
    properties:
      answers: {}
      exam_id:
        type: integer
      student_id:
        type: integer
    type: object
...
  models.Discussion:
    properties:
      description:
        type: string
      id:
        type: integer
      replies: {}
      student_id:
        type: integer
      topic:
        type: string
    type: object
  models.ExamAnswers:
    properties:
      answers: {}
      exam_id:
        type: integer
      id:
        type: integer
      status:
        type: string
      student_id:
        type: integer
    type: object
  models.ExamGrades:
    properties:
      detail: {}
      exam_id:
        type: integer
      id:
        type: integer
      score:
        type: number
      student_id:
        type: integer
    type: object
...
  models.ReplyDiscussion:
    properties:
      replies:
        type: string
      student_id:
        type: integer
      student_name:
        type: string
    type: object
...
  models.UpdateDiscussionRequest:
    properties:
      description:
        type: string
      student_id:
        type: integer
      topic:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for Project PPL - Learnify
  title: Project PPL API
  version: 0.2.1
paths:
...
  /api/v1/discussions:
    delete:
      consumes:
      - application/json
      description: Deletes a discussion from the database by ID
      parameters:
      - description: Discussion ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Discussion
      tags:
      - Discussions
    get:
      consumes:
      - application/json
      description: Fetch all discussions from the database with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 15)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Discussions
      tags:
      - Discussions
    patch:
      consumes:
      - application/json
      description: Updates an existing discussion in the database
      parameters:
      - description: Discussion ID
        in: query
        name: id
        required: true
        type: integer
      - description: Updated Discussion Data
        in: body
        name: discussion
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDiscussionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Discussion'
      security:
      - BearerAuth: []
      summary: Update Discussion
      tags:
      - Discussions
    post:
      consumes:
      - application/json
      description: Create a new discussion in the database
      parameters:
      - description: Discussion data
        in: body
        name: discussion
        required: true
        schema:
          $ref: '#/definitions/models.CreateDiscussionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Discussion'
      security:
      - BearerAuth: []
      summary: Create Discussion
      tags:
      - Discussions
  /api/v1/discussions/reply:
    patch:
      consumes:
      - application/json
      description: Replies or comments on an existing discussion
      parameters:
      - description: Discussion ID
        in: query
        name: id
        required: true
        type: integer
      - description: Updated Discussion Data
        in: body
        name: discussion
        required: true
        schema:
          $ref: '#/definitions/models.ReplyDiscussion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Discussion'
      security:
      - BearerAuth: []
      summary: Reply or comment on a Discussion
      tags:
      - Discussions
...
  /api/v1/exams-answers:
    delete:
      consumes:
      - application/json
      description: Delete an exam answer by ID
      parameters:
      - description: Exam ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Exam Answer
      tags:
      - Exam Answers (Student Answers)
    get:
      consumes:
      - application/json
      description: Fetch all exams for a specific material
      parameters:
      - description: Exam ID
        in: query
        name: exam_id
        required: true
        type: integer
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ExamAnswers'
            type: array
      security:
      - BearerAuth: []
      summary: Get Exam Answers by Exam ID
      tags:
      - Exam Answers (Student Answers)
    patch:
      consumes:
      - application/json
      description: Update an existing exam answer
      parameters:
      - description: Exam Answer ID (Student's Answer ID)
        in: query
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/models.CreateExamAnswersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExamAnswers'
      security:
      - BearerAuth: []
      summary: Update Exam Answer
      tags:
      - Exam Answers (Student Answers)
    post:
      consumes:
      - application/json
      description: Create a new exam
      parameters:
      - description: Exam answers data
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/models.CreateExamAnswersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExamAnswers'
      security:
      - BearerAuth: []
      summary: Create Exam Answers
      tags:
      - Exam Answers (Student Answers)
  /api/v1/exams/calculate-grade:
    post:
      consumes:
      - application/json
      description: Calculate grade after submitting exam
      parameters:
      - description: Exam grades data
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/models.CalculateExamGrades'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExamGrades'
      security:
      - BearerAuth: []
      summary: Calculate Grade
      tags:
      - Exam Answers (Student Answers)
  /api/v1/exams/get-all-grade:
    get:
      consumes:
      - application/json
      description: Get calculated grades for all exams
      parameters:
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExamGrades'
      security:
      - BearerAuth: []
      summary: Get ALl Grades
      tags:
      - Exam Answers (Student Answers)
  /api/v1/exams/get-grade:
    get:
      consumes:
      - application/json
      description: Get calculated grade
      parameters:
      - description: Exam ID
        in: query
        name: exam_id
        required: true
        type: integer
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExamGrades'
      security:
      - BearerAuth: []
      summary: Get Grade
      tags:
      - Exam Answers (Student Answers)
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"